requirements.txt
numpy
torch==2.3
scikit-learn
monai
pydicom
nibabel
scikit-image
scipy
pillow
gdown
ignite
tqdm
pandas
transformers
mlflow
clearml
matplotlib
tifffile
imagecodecs
pyyaml
jsonschema
pydicom
nni
ipywidgets
wandb
lightning
einops
pynrrd
lmdb
ITK
torcheval
monai-generative
pingouin
ipykernel
pytorch-ignite
scikit-multilearn
triton
xformers==0.0.26.post1